+++ src/hooks/api/use-notifications.ts
@@ -4,45 +4,21 @@
  * Provides TanStack Query hooks for notifications list + mutations via Supabase PostgREST.
  */
 
-import {
-	useMutation,
-	useQuery,
-	useQueryClient,
-	keepPreviousData,
-	mutationOptions
-} from '@tanstack/react-query'
+import { useMutation, useQuery, useQueryClient, mutationOptions } from '@tanstack/react-query'
 
 import { createClient } from '#lib/supabase/client'
 import { mutationKeys } from './mutation-keys'
-import { QUERY_CACHE_TIMES } from '#lib/constants/query-config'
 import {
 	handleMutationError,
 	handleMutationSuccess
 } from '#lib/mutation-error-handler'
+import {
+	notificationKeys,
+	notificationQueries
+} from './query-keys/notification-keys'
 import type { Database } from '#types/supabase'
 
 type NotificationItem = Database['public']['Tables']['notifications']['Row']
-
-/** Subset of notification columns for list views. */
-type NotificationListItem = Pick<NotificationItem, 'id' | 'user_id' | 'notification_type' | 'title' | 'message' | 'is_read' | 'entity_type' | 'entity_id' | 'action_url' | 'read_at' | 'created_at'>
-
-interface PaginatedNotifications {
-	data: NotificationListItem[]
-	total: number
-	page: number
-	limit: number
-}
-
-const notificationKeys = {
-	all: ['notifications'] as const,
-	list: (
-		page: number,
-		limit: number,
-		unreadOnly: boolean,
-		queryString?: string
-	) => ['notifications', { page, limit, unreadOnly, queryString }] as const,
-	unreadCount: ['notifications', 'unread-count'] as const
-}
 
 // ============================================================================
 // MUTATION OPTIONS FACTORIES
@@ -143,57 +119,11 @@ export function useNotifications(params?: {
 	limit?: number
 	unreadOnly?: boolean
 }) {
-	const page = params?.page ?? 1
-	const limit = params?.limit ?? 20
-	const unreadOnly = params?.unreadOnly ?? false
-	const queryString = new URLSearchParams({
-		page: String(page),
-		limit: String(limit),
-		...(unreadOnly ? { unreadOnly: 'true' } : {})
-	}).toString()
-
-	return useQuery({
-		queryKey: notificationKeys.list(page, limit, unreadOnly, queryString),
-		queryFn: async (): Promise<PaginatedNotifications> => {
-			const supabase = createClient()
-			let query = supabase
-				.from('notifications')
-				.select('id, user_id, notification_type, title, message, is_read, entity_type, entity_id, action_url, read_at, created_at', { count: 'exact' })
-
-			if (unreadOnly) {
-				query = query.eq('is_read', false)
-			}
-
-			const { data, count, error } = await query
-				.order('created_at', { ascending: false })
-				.range((page - 1) * limit, page * limit - 1)
-
-			if (error) throw error
-
-			return { data: data ?? [], total: count ?? 0, page, limit }
-		},
-		...QUERY_CACHE_TIMES.LIST,
-		placeholderData: keepPreviousData
-	})
+	return useQuery(notificationQueries.list(params))
 }
 
 export function useUnreadNotificationsCount() {
-	return useQuery({
-		queryKey: notificationKeys.unreadCount,
-		queryFn: async (): Promise<PaginatedNotifications> => {
-			const supabase = createClient()
-			const { count, error } = await supabase
-				.from('notifications')
-				.select('id', { count: 'exact', head: true })
-				.eq('is_read', false)
-
-			if (error) throw error
-
-			return { data: [], total: count ?? 0, page: 1, limit: 1 }
-		},
-		staleTime: 30_000,
-		gcTime: 60_000
-	})
+	return useQuery(notificationQueries.unreadCount())
 }
 
 // ============================================================================
