+++ src/hooks/api/use-payments.ts
@@ -11,18 +11,11 @@
  */
 
 import { useQuery } from '@tanstack/react-query'
-import { createClient } from '#lib/supabase/client'
-import { handlePostgrestError } from '#lib/postgrest-error-handler'
-import { QUERY_CACHE_TIMES } from '#lib/constants/query-config'
-import type { StripeSessionStatusResponse } from '#types/core'
-import type { TenantPaymentStatusResponse } from '#types/api-contracts'
-import type { SubscriptionData } from '#types/stripe'
 
 import {
 	rentCollectionQueries,
 	tenantPaymentQueries,
-	rentPaymentKeys,
-	paymentVerificationKeys
+	paymentStatusQueries
 } from './query-keys/payment-keys'
 
 // Re-export keys and query factories for consumers that import from use-payments
@@ -31,7 +24,8 @@ export {
 	rentPaymentKeys,
 	paymentVerificationKeys,
 	rentCollectionQueries,
-	tenantPaymentQueries
+	tenantPaymentQueries,
+	paymentStatusQueries
 } from './query-keys/payment-keys'
 
 // ============================================================================
@@ -67,43 +61,7 @@ export function useOverduePayments() {
  * Get current payment status for a tenant — PostgREST
  */
 export function usePaymentStatus(tenant_id: string) {
-	return useQuery({
-		queryKey: rentPaymentKeys.status(tenant_id),
-		queryFn: async (): Promise<TenantPaymentStatusResponse> => {
-			const supabase = createClient()
-			const today = new Date().toISOString().split('T')[0] as string
-			const { data, error } = await supabase
-				.from('rent_payments')
-				.select('id, amount, status, due_date, paid_date, stripe_payment_intent_id')
-				.eq('tenant_id', tenant_id)
-				.order('due_date', { ascending: false })
-				.limit(1)
-				.maybeSingle()
-			if (error) handlePostgrestError(error, 'rent_payments')
-			if (!data) {
-				return {
-					status: 'pending',
-					rent_amount: 0,
-					nextDueDate: null,
-					lastPaymentDate: null,
-					outstandingBalance: 0,
-					isOverdue: false
-				}
-			}
-			const isOverdue =
-				data.status !== 'succeeded' && data.due_date < today
-			return {
-				status: data.status as TenantPaymentStatusResponse['status'],
-				rent_amount: data.amount,
-				nextDueDate: data.due_date,
-				lastPaymentDate: data.paid_date,
-				outstandingBalance: data.status === 'succeeded' ? 0 : data.amount,
-				isOverdue
-			}
-		},
-		enabled: !!tenant_id,
-		...QUERY_CACHE_TIMES.STATS
-	})
+	return useQuery(paymentStatusQueries.status(tenant_id))
 }
 
 /**
@@ -136,24 +94,7 @@ export function usePaymentVerification(
 	sessionId: string | null,
 	options: { throwOnError?: boolean } = {}
 ) {
-	return useQuery({
-		queryKey: paymentVerificationKeys.verifySession(sessionId ?? ''),
-		queryFn: async (): Promise<{ subscription: SubscriptionData | null }> => {
-			if (!sessionId) return { subscription: null }
-			const supabase = createClient()
-			const { data, error } = await supabase.functions.invoke(
-				'stripe-checkout-session',
-				{ body: { session_id: sessionId } }
-			)
-			if (error) throw new Error(error.message ?? 'Session verification failed')
-			return { subscription: (data as { subscription: SubscriptionData | null })?.subscription ?? null }
-		},
-		enabled: !!sessionId,
-		...QUERY_CACHE_TIMES.SECURITY,
-		refetchOnReconnect: false,
-		refetchOnMount: false,
-		throwOnError: options.throwOnError ?? false
-	})
+	return useQuery(paymentStatusQueries.verifySession(sessionId, options))
 }
 
 /**
@@ -165,24 +106,5 @@ export function useSessionStatus(
 	sessionId: string | null,
 	options: { throwOnError?: boolean } = {}
 ) {
-	return useQuery({
-		queryKey: paymentVerificationKeys.sessionStatus(sessionId ?? ''),
-		queryFn: async (): Promise<StripeSessionStatusResponse> => {
-			if (!sessionId) {
-				return { status: 'expired', payment_intent_id: null, payment_status: null, payment_intent_status: null } as StripeSessionStatusResponse
-			}
-			const supabase = createClient()
-			const { data, error } = await supabase.functions.invoke(
-				'stripe-checkout-session',
-				{ body: { session_id: sessionId } }
-			)
-			if (error) throw new Error(error.message ?? 'Session status check failed')
-			return data as StripeSessionStatusResponse
-		},
-		enabled: !!sessionId,
-		...QUERY_CACHE_TIMES.STATS,
-		refetchOnReconnect: false,
-		refetchOnMount: false,
-		throwOnError: options.throwOnError ?? false
-	})
+	return useQuery(paymentStatusQueries.sessionStatus(sessionId, options))
 }
